Made 'options' an int rather than a long.
For consistency with the API.

+++ src/blocks.c
@@ -50,7 +50,7 @@ static cmark_node* make_document()
 	return e;
 }
 
-cmark_parser *cmark_parser_new(long options)
+cmark_parser *cmark_parser_new(int options)
 {
 	cmark_parser *parser = (cmark_parser*)malloc(sizeof(cmark_parser));
 	cmark_node *document = make_document();
@@ -317,7 +317,7 @@ static cmark_node* add_child(cmark_parser *parser, cmark_node* parent,
 
 // Walk through cmark_node and all children, recursively, parsing
 // string content into inline content where appropriate.
-static void process_inlines(cmark_node* root, cmark_reference_map *refmap, long options)
+static void process_inlines(cmark_node* root, cmark_reference_map *refmap, int options)
 {
 	cmark_iter *iter = cmark_iter_new(root);
 	cmark_node *cur;
@@ -422,7 +422,7 @@ static cmark_node *finalize_document(cmark_parser *parser)
 	return parser->root;
 }
 
-cmark_node *cmark_parse_file(FILE *f, long options)
+cmark_node *cmark_parse_file(FILE *f, int options)
 {
 	unsigned char buffer[4096];
 	cmark_parser *parser = cmark_parser_new(options);
@@ -442,7 +442,7 @@ cmark_node *cmark_parse_file(FILE *f, long options)
 	return document;
 }
 
-cmark_node *cmark_parse_document(const char *buffer, size_t len, long options)
+cmark_node *cmark_parse_document(const char *buffer, size_t len, int options)
 {
 	cmark_parser *parser = cmark_parser_new(options);
 	cmark_node *document;
